pyproject.toml
[project]
name = "encoder-only-transformer"
version = "0.1.0"
description = "An educational Encoder-only Transformer project built from scratch with PyTorch."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["torch", "pyyaml"]

[dependency-groups]
dev = ["pytest", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
